requirements.txt
albumentations==1.1.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
beautifultable==1.1.0
bleach==5.0.1
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028896657/work
brotlipy==0.7.0
certifi==2022.5.18.1
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dask==2021.10.0
dataclasses==0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
docker-pycreds @ file:///Users/ktietz/demo/mc3/conda-bld/docker-pycreds_1630654474270/work
docopt==0.6.2
efficientnet-pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.1
flake8 @ file:///tmp/build/80754af9/flake8_1620776156532/work
fonttools==4.25.0
fsspec @ file:///opt/conda/conda-bld/fsspec_1652802105984/work
future==0.18.2
gifmaker==1.5
gitdb==4.0.9
GitPython==3.1.27
hydra-core==1.3.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1635529103726/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
imgaug @ file:///tmp/build/80754af9/imgaug_1605226543559/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
json5==0.9.9
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
MarkupSafe==2.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
notebook==6.4.12
notebook-shim==0.1.0
numexpr @ file:///tmp/build/80754af9/numexpr_1640704205123/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
odfpy==1.4.1
olefile==0.46
omegaconf==2.3.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl @ file:///tmp/build/80754af9/openpyxl_1632777717936/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
pathtools @ file:///Users/ktietz/demo/mc3/conda-bld/pathtools_1629713893697/work
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
plotly==5.8.2
prometheus-client==0.14.1
promise @ file:///tmp/build/80754af9/promise_1614011640636/work
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes @ file:///tmp/build/80754af9/pyflakes_1617200973297/work
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-box==6.0.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1648727874099/work
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-image @ file:///tmp/build/80754af9/scikit-image_1648196043170/work
scikit-learn==1.0.2
scipy @ file:///tmp/build/80754af9/scipy_1641536880743/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
Send2Trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
Shapely @ file:///tmp/build/80754af9/shapely_1632813481411/work
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tenacity @ file:///tmp/build/80754af9/tenacity_1632746495287/work
terminado==0.15.0
threadpoolctl==3.1.0
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
timm==0.5.4
tinycss2==1.1.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
traitlets==5.3.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wandb==0.12.20
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
XlsxWriter @ file:///opt/conda/conda-bld/xlsxwriter_1649073856329/work
yarg==0.1.9
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
